wscript: Rename lwrcase_dirname (with values like "ardourX") to INTERNAL_NAME
The build system does a lot all over to keep ardour8 (almost) without
file name conflicts if installed next to ardour9. Often by appending
MAJOR in lots of places.

lwrcase_dirname was just 'ardour' + MAJOR. The naming of lwrcase_dirname
did not say much about what it was, and it was only defined locally in
build(). We set INTERNAL_NAME globally in configure(), exactly like
MAJOR is set.

INTERNAL_NAME is similar to the existing PROGRAM_NAME, which contains
the visible name. INTERNAL_NAME is however more intended to be used for
namespacing installed files. We will soon use it in a lot of places
where MAJOR used to be used.

The end goal is to make sure there only is one place where the
"ardour"+major naming policy is hardcoded. And to make it configurable
instead of hardcoded.

Things should not be repeated so they are hard to change, and they
should not be hardcoded. The policy of including the major version in
the internal name is not necessarily relevant when always using latest
version. Most programs don't do that. But also, if there for some reason
should be a need to install several minor versions side by side, then it
would be relevant to include both the major and the minor number.