surface us2400: fix bad button configuration
Tests showed
	ControlSurfacesTest: Tascam US-2400
	: [ERROR]: Unknown button ID "F7"
	: [ERROR]: Unknown button ID "F8"

This is trying to mitigate it blindfolded.

It doesn't seem like Tascam US-2400 has more than 6 buttons, so it seems
like the profile has been incorrect since it was introduced in 937efdb6.

(But as pointed out in e6de7957, it seems like the profile often isn't
installed and button assignment thus isn't used.)