temporal: rename set_sample_rate to set_most_recent_engine_sample_rate to clarify uee
Avoid ambiguity with other uses of set_sample_rate. Make it clear that
it only is invoked in one place.

This is a global variable similar to the superclock rate, and it is
crucial that values always are read from sessions, and set back to the
right default when creating new sessions.