temporal: drop hardcoded TempoMap defaults
9964f20c just introduced proper initialization of the TempoMap whenever
a new session is created. For existing sessions, the TempoMap will
already be initialized from the session.

There is thus no longer any need for setting defaults it in the early
low level initialization in TempoMap::init . It is much better to avoid
hardcoding such "policy", and instead set it at a higher level.

A TempoMap without defaults will have limited use, but it will fail
quite clearly if there should be code paths that fail to initialize it.

When loading sessions, this change will stop uninitialized use of
superclock_ticks_per_second(), and DEBUG_EARLY_SCTS_USE should thus in
theory work for that code path.